+++ packages/component-library/src/components/plmg-dropdown-item/test/plmg-dropdown-item.spec.tsx
@@ -14,7 +14,7 @@ describe('plmg-dropdown-item', () => {
         <span class="plmg-dropdown-item-content">
           Dropdown Item
         </span>
-        <plmg-svg-icon class="plmg-dropdown-item-active-icon" icon="check"></plmg-svg-icon>
+        <plmg-svg-icon class="plmg-dropdown-item-selected-icon" icon="check"></plmg-svg-icon>
         </a>
       </mock:shadow-root>
     </plmg-dropdown-item>
@@ -33,7 +33,26 @@ describe('plmg-dropdown-item', () => {
         <span class="plmg-dropdown-item-content">
           Dropdown Item
         </span>
-        <plmg-svg-icon class="plmg-dropdown-item-active-icon" icon="check"></plmg-svg-icon>
+        <plmg-svg-icon class="plmg-dropdown-item-selected-icon" icon="check"></plmg-svg-icon>
+        </a>
+      </mock:shadow-root>
+    </plmg-dropdown-item>
+    `);
+  });
+  it('renders with selected classes', async () => {
+    const page = await newSpecPage({
+      components: [PlmgDropdownItem],
+      html: `<plmg-dropdown-item icon="duck" text='Dropdown Item' href='https://ducky.eco' selected rel='noopener noreferrer' target='_blank'></plmg-dropdown-item>`,
+    });
+    expect(page.root).toEqualHtml(`
+    <plmg-dropdown-item href="https://ducky.eco" icon="duck" selected rel="noopener noreferrer" target="_blank" text="Dropdown Item">
+      <mock:shadow-root>
+      <a class="plmg-dropdown-item plmg-dropdown-item-selected" href="https://ducky.eco" rel="noopener noreferrer" title="Dropdown Item">
+      <plmg-svg-icon class="plmg-dropdown-item-icon" icon="duck"></plmg-svg-icon>
+        <span class="plmg-dropdown-item-content">
+          Dropdown Item
+        </span>
+        <plmg-svg-icon class="plmg-dropdown-item-selected-icon plmg-dropdown-item-selected-icon-visible" icon="check"></plmg-svg-icon>
         </a>
       </mock:shadow-root>
     </plmg-dropdown-item>